requirements.txt
grpcio==1.48.1
grpcio-tools==1.48.1
Pillow==9.2.0
protobuf==3.20.2
six==1.16.0
